+++ packages/commonwealth/client/scripts/views/components/ThreadContestTag/index.ts
@@ -1,3 +1,5 @@
 import ThreadContestTag from './ThreadContestTag';
+import ThreadContestTagContainer from './ThreadContestTagContainer';
 
+export { ThreadContestTagContainer };
 export default ThreadContestTag;
