+++ packages/commonwealth/client/scripts/views/components/component_kit/CWContentPage/CWContentPage.tsx
@@ -1,6 +1,5 @@
 import { getThreadActionTooltipText } from 'helpers/threads';
 import { truncate } from 'helpers/truncate';
-import { useFlag } from 'hooks/useFlag';
 import useUserActiveAccount from 'hooks/useUserActiveAccount';
 import { IThreadCollaborator } from 'models/Thread';
 import moment from 'moment';
@@ -10,8 +9,7 @@ import { useSearchParams } from 'react-router-dom';
 import app from 'state';
 import { useRefreshMembershipQuery } from 'state/api/groups';
 import Permissions from 'utils/Permissions';
-import ThreadContestTag from 'views/components/ThreadContestTag';
-import { getWinnersFromAssociatedContests } from 'views/components/ThreadContestTag/utils';
+import { ThreadContestTagContainer } from 'views/components/ThreadContestTag';
 import { isHot } from 'views/pages/discussions/helpers';
 import Account from '../../../../models/Account';
 import AddressInfo from '../../../../models/AddressInfo';
@@ -125,7 +123,6 @@ export const CWContentPage = ({
   const [urlQueryParams] = useSearchParams();
   const { activeAccount: hasJoinedCommunity } = useUserActiveAccount();
   const [isUpvoteDrawerOpen, setIsUpvoteDrawerOpen] = useState<boolean>(false);
-  const contestsEnabled = useFlag('contest');
 
   const { data: memberships = [] } = useRefreshMembershipQuery({
     communityId: app.activeChainId(),
@@ -225,32 +222,14 @@ export const CWContentPage = ({
     isThreadTopicGated: isRestrictedMembership,
   });
 
-  const contestWinners = getWinnersFromAssociatedContests(
-    thread?.associatedContests,
-  );
-  const showContestWinnerTag = contestsEnabled && contestWinners.length > 0;
-
   const mainBody = (
     <div className="main-body-container">
       <div className="header">
         {typeof title === 'string' ? (
           <h1 className="title">
-            {showContestWinnerTag &&
-              contestWinners.map((winner, index) => {
-                if (!winner) {
-                  return null;
-                }
-
-                return (
-                  <ThreadContestTag
-                    date={winner.date}
-                    round={winner.round}
-                    title={winner.title}
-                    prize={winner.prize}
-                    key={index}
-                  />
-                );
-              })}
+            <ThreadContestTagContainer
+              associatedContests={thread?.associatedContests}
+            />
             {truncate(title)}
           </h1>
         ) : (
